FIX : Minor cleanup to IDb with sub interfaces
