+++ src/lib/kotlin/slatekit-common/src/main/kotlin/slatekit/common/db/ProcSupport.kt
@@ -1,0 +1,42 @@
+package slatekit.common.db
+
+import java.sql.ResultSet
+
+interface ProcSupport {
+
+    /**
+     * Calls a stored procedure
+     * @param procName : The name of the stored procedure e.g. get_by_id
+     * @param callback : The callback to handle the resultset
+     * @param moveNext : Whether or not to automatically move the resultset to the next/first row
+     * @param inputs : The parameters for the stored proc. The types will be auto-converted my-sql types.
+     */
+    fun <T> callQuery(
+            procName: String,
+            callback: (ResultSet) -> T?,
+            moveNext: Boolean = true,
+            inputs: List<Any>? = null
+    ): T?
+
+    /**
+     * Calls a stored procedure
+     * @param procName : The name of the stored procedure e.g. get_by_id
+     * @param callback : The callback to handle the resultset
+     * @param moveNext : Whether or not to automatically move the resultset to the next/first row
+     * @param inputs : The parameters for the stored proc. The types will be auto-converted my-sql types.
+     */
+    fun <T> callQueryMapped(
+            procName: String,
+            mapper: Mapper,
+            inputs: List<Any>? = null
+    ): List<T>?
+
+    /**
+     * Calls a stored procedure
+     * @param procName : The name of the stored procedure e.g. get_by_id
+     * @param callback : The callback to handle the resultset
+     * @param moveNext : Whether or not to automatically move the resultset to the next/first row
+     * @param inputs : The parameters for the stored proc. The types will be auto-converted my-sql types.
+     */
+    fun callUpdate(procName: String, inputs: List<Any>? = null): Int
+}
